refactor: 프로필 수정 컴포넌트 비즈니스 로직 분리
