+++ src/Pages/MyPage/EditProfile/EditProfile.tsx
@@ -1,82 +1,22 @@
-import { useEffect, useRef, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import defaultProfileImg from '@Assets/Images/default_profile.svg';
 import ProfileSetupForm from '@components/common/ProfileSetupForm';
 import ConfirmModal from '@components/common/ConfirmModal';
 import useModal from '@hooks/useModal';
 import { useUser } from '@hooks/useUser';
-import { useUpdateUserInfoMutation } from '@store/api/userApiSlice';
 import Spinner from '@components/Auth/Spinner';
-import { useDispatch, useSelector } from 'react-redux';
-import {
-  resetState,
-  setLatitude,
-  setLongitude,
-  setRegionCode,
-} from '@store/slices/coordinateSlice';
-import { RootState } from '@store/types';
-import { IUpdateUser } from '@/types/UserTypes';
+import { useEditProfile } from '@hooks/useEditProfile';
 
 function EditProfile() {
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const { regionCode, latitude, longitude } = useSelector(
-    (state: RootState) => state.coordinateSetup
-  );
   const {
     data: { data: user },
   } = useUser();
-  const initialNickname = useRef(user?.nickname);
-  const newNickname = useRef(user?.nickname);
-  const imgFile = useRef<File | null>(null);
   const { isOpen, open, close } = useModal();
-  const [updateUser, { isLoading: isUpdateUserLoading }] =
-    useUpdateUserInfoMutation();
-
-  const handleSubmit = async (nickname: string, newImgFile: File) => {
-    newNickname.current = nickname;
-    imgFile.current = newImgFile;
-    open();
-  };
-
-  const handleUpdateUser = async () => {
-    const formData: IUpdateUser = {
-      request: {
-        addressRequest: {
-          regionCode,
-          latitude,
-          longitude,
-          type: 'main',
-        },
-      },
-    };
-    if (initialNickname.current !== newNickname.current) {
-      formData.request.nickname = newNickname.current;
-    }
-    if (imgFile.current) {
-      formData.file = imgFile.current;
-    }
-
-    try {
-      await updateUser(formData).unwrap();
-    } catch (error: any) {
-      throw new Error(error.data.message);
-    }
-  };
-
-  const onModalCompleteButtonClick = () => {
-    navigate('/my-page');
-  };
-
-  useEffect(() => {
-    dispatch(setRegionCode(user?.address.hdongCode));
-    dispatch(setLatitude(user?.address.latitude));
-    dispatch(setLongitude(user?.address.longitude));
-
-    return () => {
-      dispatch(resetState());
-    };
-  }, []);
+  const {
+    handleSubmit,
+    handleUpdateUser,
+    onModalCompleteButtonClick,
+    isLoading,
+  } = useEditProfile(user, open);
 
   return (
     <>
@@ -96,7 +36,7 @@ function EditProfile() {
         onCompletedAction={onModalCompleteButtonClick}
         closeAction={close}
       />
-      {isUpdateUserLoading && <Spinner />}
+      {isLoading && <Spinner />}
     </>
   );
 }
